Copy distinct list when merging
This way we no longer modify the references of the items in the list.

We also prevent duplicate events by converting to a set and then back to a list sorted by time. This may not be exactly how the input was sorted in rare cases, but it's within acceptable bounds.