Replace `had_value` with `yielded`
This way if we get events which we never yield, it will eventually return an error (e.g. events after a certain page all being the same).

The program will only look for events as old as when it starts searching, so the chances of us finding 10 pages of newer events is very small.