Add nomination comment to populator
It is much more fitting to include this in the database rather than retrieving it at the time of formatting the event.

This way there'd be no need for time-sensitive retrieval of events *before* some nomination, if retrieved at a later date. Additionally, you'd be able to filter by content for these comments.