Reduce strictness of no event yielding
Now requires that no events are generated from the page, as opposed to no new events within the timeframe being found.

An infinite loop on the same page could be caused by the latest event being moved over to the previous page while we are going to the next page (i.e. all events are too new).