Remove HTTPException error handling
HTTPException is an umbrella class for all HTTP-related exceptions given by the discord api.

As such we do not know the circumstance which caused the error to raise, meaning it is difficult to properly tailor an appropriate response.

Instead of logging, we should look at each error individually and create additional except statements if it's something we can safely log.