Move news embed field value to footer
As a result we also need to change our field title and a regular title, so that we can have no description under it.

The reason behind this change is balancing out the visual aspects of the embed so that the image is more in the middle rather than at the bottom.