Add `CachedDatabase`
This is much faster than a regular database in cases where we commonly request the same data, for example in statistical analysis.

Drawback of this being that the cache must be cleared in order for the same request to retrieve different data (e.g. in case an entry is updated).