Add parsing support for genre/language edits
Uses the content field like "{old} -> {new}", for a lack of a better place to put this (making new db fields for this seemed a bit overkill).

Despite both new and old being in a single field, they can still be filtered for using e.g. "% -> Electronic" for anything changed to "Electronic".