Rework removal reason test to new last eval
bnsite now has a "latestEvaluation" route, which we should use instead of "bnRemoval" due to resignations otherwise not working properly.

Also makes it close to impossible to do card archiving/group changes in the wrong order, since we can just grab the consensus from the card whether archived or not.