Memoize `split_unescaped`
Total time spent in this function was previously 40 seconds in ~5 min of profiling. Now it's 1 second.

Since subscriptions don't change often, and many are the same, caching the result of this method turned out to be pretty effective.

The delimiter argument must be hashable, in order to work with the memoized decorator, so it's now a tuple.