Add allow_api arg to Beatmapset
This is useful when testing, as it allows tests focusing on single properties of the beatmapset to not need to set anything else.

Previously when we added an argument to either User or Beatmapset, we needed to add this to all tests as well. The less changes cascade into other changes, the better. So this fixes that.